Do global search/replace for customize control selector
In WP_Customize_Conrol::render(), the $id for a control is constructed via:

  $id    = 'customize-control-' . str_replace( '[', '-', str_replace( ']', '', $this->id ) );

The PHP str_replace function does a global replacement. In JavaScript, however, if you
call the string.replace() with a string as an argument instead of a RegExp, it will do
a single replacement instead of a global one. Therefore, RegExp literals with the g flag
must be supplied to ensure that the ID calculated in JS matches the ID calculated in PHP.